# api/swagger.yaml
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: acgw_session
  schemas:
...
    AdminDiscordRoleMapping:
      properties:
        discord_role_id:
          type: string
        role:
          type: string
      type: object
...
    AdminPermissionRequest:
      properties:
        permission:
          type: string
      type: object
    AdminRoleGrant:
      properties:
        permission:
          type: string
        role:
          type: string
      type: object
    AdminRolesResponse:
      properties:
        grants:
          items:
            $ref: '#/components/schemas/AdminRoleGrant'
          type: array
          uniqueItems: false
        mappings:
          items:
            $ref: '#/components/schemas/AdminDiscordRoleMapping'
          type: array
          uniqueItems: false
      type: object
...
    ErrorBody:
      properties:
        code:
          type: string
        details: {}
        message:
          type: string
      type: object
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        request_id:
          type: string
      type: object
...
    ListUsersResponse:
      properties:
        users:
          items:
            $ref: '#/components/schemas/User'
          type: array
          uniqueItems: false
      type: object
...
    User:
      properties:
        created_at:
          type: string
        discord_id:
          type: string
        display_name:
          type: string
        global_name:
          type: string
        user_id:
          type: string
        username:
          type: string
      type: object
...
info:
  description: |-
    REST gateway between community applications and AzerothCore.
    It hides the AzerothCore SOAP/CLI internals behind a typed, authenticated and authorized HTTP API.
  title: ac-community-gw API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1/admin/roles:
    get:
      description: Returns every role -> permission grant and Discord role mapping.
        Requires the gw.identity.roles.manage permission.
      operationId: identity.admin.roles.list
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRolesResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: List roles and mappings
      tags:
      - identity
  /api/v1/admin/roles/{role}/permissions:
    post:
      description: Grants a permission to a role (creating the role if needed). Requires
        the gw.identity.roles.manage permission.
      operationId: identity.admin.roles.grant
      parameters:
      - description: role
        in: path
        name: role
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPermissionRequest'
        description: permission
        required: true
      responses:
        "204":
          description: Granted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Grant a permission to a role
      tags:
      - identity
  /api/v1/admin/roles/{role}/permissions/{permission}:
    delete:
      description: Revokes a permission from a role. Requires the gw.identity.roles.manage
        permission.
      operationId: identity.admin.roles.revoke
      parameters:
      - description: role
        in: path
        name: role
        required: true
        schema:
          type: string
      - description: permission
        in: path
        name: permission
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Revoked
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Revoke a permission from a role
      tags:
      - identity
...
  /api/v1/identity/users:
    get:
      description: Lists provisioned community users. Requires the gw.identity.user.read
        permission.
      operationId: identity.users.list
      parameters:
      - description: filter by username or Discord name
        in: query
        name: filter
        schema:
          type: string
      - description: page size
        in: query
        name: limit
        schema:
          default: 50
          type: integer
      - description: page offset
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: List community users
      tags:
      - identity
...